+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for `lowpass-filter`
 
+## v0.4.0 (2025-07-06)
+- modernized crate, Rust edition 2024
+- Added new `LowpassFilter` type that makes it easier to use this crate
+- MSRV is now 1.85.0
+
 ## v0.3.1/0.3.2 (2021-11-15)
 - smaller crate size/don't include irrelevant stuff